requirements.txt
python-dotenv==1.0.0
PySide6
libsonic-d
requests==2.31.0
Flask==2.3.3
Flask-SocketIO==5.3.6
gunicorn==21.2.0
Werkzeug==2.3.7
gevent==23.9.1
python-engineio==4.8.0
python-socketio==5.10.0
loguru==0.7.2
simplejson==3.19.2
pyttsx3==2.90
pywin32==306 ; sys_platform == 'win32'
